requirements.txt
fastapi
uvicorn[standard]
pydantic
langgraph
langchain-groq
docker
tree-sitter
tree-sitter-python
python-dotenv
